requirements.txt
flask==3.0.2
pymongo[srv]==4.6.2
gunicorn
python-dotenv